// projects/adminmodule/src/app/models/responsemodel.ts
export interface AdmissionQuotasubjectGroups {
    Batch_code: number
    Subject_group_id: number
    Subject_group_code: string
    Subject_group_name: string
    Quota_status: string
}
// ...
export interface Subject_group_nep {
    Batch_code: number
    Batch_name: string
    Subject_group_code: string
    Major: string
    Otherlevel: string
    Otherlevelcode: number
    Levelno: number
    Levelnomenclature: string
    Message: string
    Quota_status: string
}